+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ### Changed 
 
 - Added the discovery type configurable, with a default value of bpnl in (ConnectorEndpointsService) (#12)
+- secured endpoints /ess/notification/receive-recursive and /ess/notification/receive with api key
 
 ## [5.4.1] - 2024-08-19